+++ e2e/commands/untrack.e2e.js
@@ -35,6 +35,14 @@ describe('bit untrack command', function () {
       expect(Object.keys(bitMap)).to.be.ofSize(1);
       expect(bitMap).to.have.property('bar/foo2');
     });
+    it('Should resolve and untrack  component and add global as prefix component ', () => {
+      helper.createComponent('bar', 'foo.js');
+      helper.addComponentWithOptions(path.normalize('bar/foo.js'), { i: 'bar' });
+      helper.untrackComponent('bar');
+      const bitMap = helper.readBitMap();
+      expect(Object.keys(bitMap)).to.be.ofSize(0);
+    });
+
     it('Should remove 2 new components and keep commited component', () => {
       helper.createComponent('bar', 'foo.js');
       helper.addComponentWithOptions(path.normalize('bar/foo.js'), { i: 'bar/foo' });
@@ -60,6 +68,13 @@ describe('bit untrack command', function () {
       const bitMap = helper.readBitMap();
       expect(Object.keys(bitMap)).to.be.ofSize(1);
       expect(bitMap).to.have.property('bar/foo2');
+    });
+    it('Should not show component if bit.json is corrupted', () => {
+      helper.corruptBitJson();
+      const untrackCmd = () => helper.untrackComponent();
+      expect(untrackCmd).to.throw(
+        'error: invalid bit.json: SyntaxError: Unexpected token o in JSON at position 1 is not a valid JSON file.'
+      );
     });
   });
 });