+++ src/consumer/component/consumer-component.js
@@ -14,6 +14,7 @@ import docsParser, { Doclet } from '../../jsdoc/parser';
 import specsRunner from '../../specs-runner';
 import SpecsResults from '../specs-results';
 import ComponentSpecsFailed from '../exceptions/component-specs-failed';
+import MissingFilesFromComponent from './exceptions/missing-files-from-component';
 import ComponentNotFoundInPath from './exceptions/component-not-found-in-path';
 import IsolatedEnvironment from '../../environment';
 import type { Log } from '../../scope/models/version';
@@ -504,11 +505,19 @@ export default class Component {
 
           return this.specsResults;
         } catch (err) {
-          if (rejectOnFailure) {
-            if (verbose) throw err;
-            throw new ComponentSpecsFailed();
-          }
-          return this.specsResults;
+          // Put this condition in comment for now because we want to catch exceptions in the testers
+          // We can just pass the rejectOnFailure=true in the consumer.runComponentSpecs
+          // Because this will also affect the condition few lines above:
+          // if (rejectOnFailure && !this.specsResults.every(element => element.pass)) {
+          // in general there is some coupling with the test running between the params:
+          // rejectOnFailure / verbose and the initiator of the running (scope / consumer)
+          // We should make a better architecture for this
+
+          // if (rejectOnFailure) {
+          if (verbose) throw err;
+          throw new ComponentSpecsFailed();
+          // }
+          // return this.specsResults;
         }
       };
 
@@ -752,9 +761,7 @@ export default class Component {
         }
       });
       if (filesKeysToDelete.length && !sourceFiles.length) {
-        throw new Error(
-          `invalid component ${id}, all files were deleted, please remove the component using bit remove command`
-        );
+        throw new MissingFilesFromComponent(id.toString());
       }
       if (filesKeysToDelete.length) {
         filesKeysToDelete.forEach(key => files.splice(key, 1));