+++ src/utils/index.js
@@ -66,6 +66,7 @@ import getExt from './fs/get-ext';
 import outputJsonFile from './fs-output-json-sync';
 import searchFilesIgnoreExt from './fs/search-files-ignore-ext';
 import { pathNormalizeToLinux, pathJoinLinux, pathRelative, pathResolve, pathJoinOs } from './path';
+import getMissingTestFiles from './getMissingTestFiles';
 
 export {
   parseSSHUrl,
@@ -139,5 +140,6 @@ export {
   pathResolve,
   pathJoinOs,
   outputJsonFile,
-  searchFilesIgnoreExt
+  searchFilesIgnoreExt,
+  getMissingTestFiles
 };